+++ src/js/Line.d.ts
@@ -11,7 +11,8 @@
  * @modified 2019-12-15 Added the Line.moveTo(Vertex) function.
  * @modified 2020-03-16 The Line.angle(Line) parameter is now optional. The baseline (x-axis) will be used if not defined.
  * @modified 2020-03-23 Ported to Typescript from JS.
- * @version  2.1.2
+ * @modified 2020-12-04 The `intersection` function returns undefined if both lines are parallel.
+ * @version  2.1.3
  *
  * @file Line
  * @public
@@ -46,11 +47,11 @@ export declare class Line extends VertTuple<Line> implements SVGSerializable {
      *
      * @method intersection
      * @param {Line} line The second line.
-     * @return {Vertex} The intersection (may lie outside the end-points).
+     * @return {Vertex|undefined} The intersection (may lie outside the end-points) or `undefined` if both lines are parallel.
      * @instance
      * @memberof Line
      **/
-    intersection(line: Line): Vertex;
+    intersection(line: Line): Vertex | undefined;
     /**
      * Create an SVG representation of this line.
      *